Refactor: Header, useTranslation
Changed to .tsx
