+++ src/app/utils/useTranslation.tsx
@@ -1,13 +1,17 @@
-'use client'
+'use client';
 import { useState, useEffect } from "react";
 
-export default function useTranslation(lang) {
-  const [translations, setTranslations] = useState({});
+interface Translations {
+  [key: string]: string;
+}
+
+export default function useTranslation(lang: string): Translations {
+  const [translations, setTranslations] = useState<Translations>({});
 
   useEffect(() => {
     const loadTranslations = async () => {
       const response = await fetch(`/locales/${lang}.json`);
-      const data = await response.json();
+      const data: Translations = await response.json();
       setTranslations(data);
     };
     loadTranslations();
